requirements.txt
django
python-dotenv
psycopg2
pillow
celery
flower
WeasyPrint
redis
djangorestframework
django-filter
social-auth-app-django
django-extensions
Werkzeug
pyOpenSSL
